Remove entreprise and book feature, add SellingBook feature
